app/svc/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-pptx==0.6.23
pdfplumber==0.10.3
pdfminer.six==20221105
pillow==10.1.0
pytesseract==0.3.10
pydantic==2.5.0
numpy==1.26.2
python-multipart==0.0.6
aiofiles==23.2.1
pytest==7.4.3
httpx==0.25.2
reportlab==4.0.7